fix: fix loses integer precision warning

+++ Sources/Amplitude/Amplitude.m
@@ -580,7 +580,7 @@ static NSString *const SEQUENCE_NUMBER = @"sequence_number";
 
         [self truncateEventQueues];
 
-        int eventCount = [self->_eventsBuffer count] + [self->_identifyBuffer count]; // refetch since events may have been deleted
+        long eventCount = [self->_eventsBuffer count] + [self->_identifyBuffer count]; // refetch since events may have been deleted
         if ((eventCount % self.eventUploadThreshold) == 0 && eventCount >= self.eventUploadThreshold) {
             [self uploadEvents];
         } else {
@@ -591,11 +591,11 @@ static NSString *const SEQUENCE_NUMBER = @"sequence_number";
 
 - (void)truncateEventQueues {
     int numEventsToRemove = MIN(MAX(1, self.eventMaxCount/10), kAMPEventRemoveBatchSize);
-    int eventCount = [self->_eventsBuffer count];
+    long eventCount = [self->_eventsBuffer count];
     if (eventCount > self.eventMaxCount) {
         self->_eventsBuffer = [[self->_eventsBuffer subarrayWithRange:NSMakeRange(0, eventCount - numEventsToRemove)] mutableCopy];
     }
-    int identifyCount = [self->_identifyBuffer count];
+    long identifyCount = [self->_identifyBuffer count];
     if (identifyCount > self.eventMaxCount) {
         self->_identifyBuffer = [[self->_identifyBuffer subarrayWithRange:NSMakeRange(0, identifyCount - numEventsToRemove)] mutableCopy];
     }
@@ -762,16 +762,16 @@ static NSString *const SEQUENCE_NUMBER = @"sequence_number";
             return;
         }
 
-        int eventCount = [self->_eventsBuffer count];
-        int numEvents = limit > 0 ? fminl(eventCount, limit) : eventCount;
+        long eventCount = [self->_eventsBuffer count];
+        long numEvents = limit > 0 ? fminl(eventCount, limit) : eventCount;
         if (numEvents == 0) {
             self->_updatingCurrently = NO;
             [self endBackgroundTaskIfNeeded];
             return;
         }
         
-        int identifyCount = [self->_identifyBuffer count];
-        int numIdentify = limit > 0 ? fminl(identifyCount, limit) : identifyCount;
+        long identifyCount = [self->_identifyBuffer count];
+        long numIdentify = limit > 0 ? fminl(identifyCount, limit) : identifyCount;
         
         NSMutableArray *events = [[self->_eventsBuffer subarrayWithRange:NSMakeRange(0, numEvents)] mutableCopy];
         NSMutableArray *identifys = [[self->_identifyBuffer subarrayWithRange:NSMakeRange(0, numIdentify)] mutableCopy];
